fix: Optimize Single->Half conversion and fix mean axis dtype
- Use direct float->Half cast instead of going through double
- Keep axis mean output as Double for compatibility with axis reduction kernels
- Element-wise mean still preserves dtype per NumPy 2.x (float32->float32)
- Update test to reflect element-wise mean dtype preservation

+++ src/NumSharp.Core/Utilities/Converts.cs
@@ -484,7 +484,7 @@ namespace NumSharp.Utilities
                         case NPTypeCode.UInt64:  return (Half)(double)Unsafe.As<T, ulong>(ref value);
                         case NPTypeCode.Char:    return (Half)(double)Unsafe.As<T, char>(ref value);
                         case NPTypeCode.Double:  return (Half)Unsafe.As<T, double>(ref value);
-                        case NPTypeCode.Single:  return (Half)(double)Unsafe.As<T, float>(ref value);
+                        case NPTypeCode.Single:  return (Half)Unsafe.As<T, float>(ref value);
                         case NPTypeCode.Decimal: return (Half)(double)Unsafe.As<T, decimal>(ref value);
                         default:
                             throw new NotSupportedException();
